my-flask-project/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
click-default-group==1.2.4
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
filelock==3.12.4
Flask==2.3.3
Flask-Session==0.5.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
google-ai-generativelanguage==0.3.2
google-api-core==2.12.0
google-auth==2.23.0
google-generativeai==0.2.0
googleapis-common-protos==1.60.0
googletrans==3.0.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.17.2
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
libretranslatepy==2.1.1
llm==0.11
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.0
patsy==0.5.3
Pillow==9.5.0
plotly==5.17.0
plotly-express==0.4.1
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-ulid==2.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
replicate==0.13.0
requests==2.31.0
rfc3986==1.5.0
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scipy==1.11.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlite-fts4==1.0.3
sqlite-migrate==0.1a2
sqlite-utils==3.35.1
statsmodels==0.14.0
streamlit==1.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.2
translate==3.6.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
